Move getIsHosted and getTokenFeature to metabase/selectors/settings
