+++ enterprise/frontend/src/metabase-enterprise/license/use-upsell-flow.ts
@@ -4,8 +4,8 @@ import { t } from "ttag";
 import { useUpsellLink } from "metabase/common/components/upsells/components/use-upsell-link";
 import { useSetting, useStoreUrl, useToast } from "metabase/common/hooks";
 import { useSelector } from "metabase/redux";
+import { getIsHosted } from "metabase/selectors/settings";
 import { getUser } from "metabase/selectors/user";
-import { getIsHosted } from "metabase/setup/selectors";
 import { useLicense } from "metabase-enterprise/settings/hooks/use-license";
 
 const NOTIFICATION_TIMEOUT = 30_000;
