+++ enterprise/frontend/src/metabase-enterprise/transforms-python/upsells/PythonTransformsUpsellModal/TransformInspectorUpsellPage.tsx
@@ -2,9 +2,9 @@ import { LoadingAndErrorWrapper } from "metabase/common/components/LoadingAndErr
 import { DottedBackground } from "metabase/common/components/upsells/components/DottedBackground";
 import { PageContainer } from "metabase/data-studio/common/components/PageContainer";
 import { useSelector } from "metabase/redux/hooks";
+import { getIsHosted } from "metabase/selectors/settings";
 import { getStoreUsers } from "metabase/selectors/store-users";
 import { getUserIsAdmin } from "metabase/selectors/user";
-import { getIsHosted } from "metabase/setup/selectors";
 import { TransformHeader } from "metabase/transforms/components/TransformHeader";
 import { useTransformWithPolling } from "metabase/transforms/hooks/use-transform-with-polling";
 import { Card, Center, Stack } from "metabase/ui";
