+++ frontend/src/embedding-sdk-bundle/hooks/private/use-sdk-usage-problem.ts
@@ -13,7 +13,7 @@ import type { MetabaseAuthConfig } from "embedding-sdk-bundle/types/auth-config"
 import { useSetting } from "metabase/common/hooks";
 import { EMBEDDING_SDK_CONFIG } from "metabase/embedding-sdk/config";
 import type { MetabaseEmbeddingSessionToken } from "metabase/embedding-sdk/types/refresh-token";
-import { getTokenFeature } from "metabase/setup/selectors";
+import { getTokenFeature } from "metabase/selectors/settings";
 
 export function useSdkUsageProblem({
   authConfig,
