+++ frontend/src/embedding-sdk-bundle/store/selectors.ts
@@ -1,8 +1,7 @@
 import type { SdkStoreState } from "embedding-sdk-bundle/store/types";
 import { EMBEDDING_SDK_CONFIG } from "metabase/embedding-sdk/config";
 import type { State } from "metabase/redux/store";
-import { getSetting } from "metabase/selectors/settings";
-import { getTokenFeature } from "metabase/setup";
+import { getSetting, getTokenFeature } from "metabase/selectors/settings";
 
 export const getIsGuestEmbedRaw = (state: SdkStoreState) =>
   state.sdk?.isGuestEmbed;
