+++ frontend/src/metabase/admin/permissions/selectors/data-permissions/permission-editor.tsx
@@ -17,8 +17,7 @@ import {
 } from "metabase/plugins";
 import type { State } from "metabase/redux/store";
 import { getMetadataWithHiddenTables } from "metabase/selectors/metadata";
-import { getSetting } from "metabase/selectors/settings";
-import { getTokenFeature } from "metabase/setup";
+import { getSetting, getTokenFeature } from "metabase/selectors/settings";
 import { getResponseErrorMessage } from "metabase/utils/errors";
 import type Schema from "metabase-lib/v1/metadata/Schema";
 import type {
