+++ frontend/src/metabase/admin/upsells/UpsellHosting.tsx
@@ -4,7 +4,7 @@ import { UpsellBanner } from "metabase/common/components/upsells/components";
 import { useSetting } from "metabase/common/hooks";
 import { getPlan, isProPlan } from "metabase/common/utils/plan";
 import { useSelector } from "metabase/redux";
-import { getIsHosted } from "metabase/setup/selectors";
+import { getIsHosted } from "metabase/selectors/settings";
 
 export const UpsellHostingBanner = ({ location }: { location: string }) => {
   const isHosted = useSelector(getIsHosted);
