+++ frontend/src/metabase/common/components/AppBanner/AppBanner.tsx
@@ -11,8 +11,8 @@ import { ReadOnlyBanner } from "metabase/nav/components/ReadOnlyBanner";
 import { TrialBanner } from "metabase/nav/components/TrialBanner";
 import { PLUGIN_SECURITY_CENTER } from "metabase/plugins";
 import { useSelector } from "metabase/redux";
+import { getIsHosted } from "metabase/selectors/settings";
 import { getUserIsAdmin } from "metabase/selectors/user";
-import { getIsHosted } from "metabase/setup/selectors";
 import { isWithinIframe } from "metabase/utils/iframe";
 
 import { getCurrentUTCTimestamp, shouldShowTrialBanner } from "./utils";
