+++ frontend/src/metabase/common/components/upsells/components/UpsellCardContent.tsx
@@ -9,9 +9,9 @@ import {
   trackUpsellViewed,
 } from "metabase/common/components/upsells/components/analytics";
 import { useSelector } from "metabase/redux";
+import { getIsHosted } from "metabase/selectors/settings";
 import { getStoreUsers } from "metabase/selectors/store-users";
 import { getUserIsAdmin } from "metabase/selectors/user";
-import { getIsHosted } from "metabase/setup/selectors";
 import {
   Box,
   Card,
