+++ frontend/src/metabase/common/hooks/use-has-token-feature/use-has-token-feature.ts
@@ -1,5 +1,5 @@
 import { useSelector } from "metabase/redux";
-import { getTokenFeature } from "metabase/setup/selectors";
+import { getTokenFeature } from "metabase/selectors/settings";
 import type { TokenFeature } from "metabase-types/api";
 
 export const useHasTokenFeature = (settingName: TokenFeature) => {
