+++ frontend/src/metabase/databases/components/DatabaseFormError/useTroubleshootingTips.tsx
@@ -5,9 +5,8 @@ import { identity } from "underscore";
 
 import { useSelector } from "metabase/redux";
 import type { State } from "metabase/redux/store";
-import { getDocsUrl } from "metabase/selectors/settings";
+import { getDocsUrl, getIsHosted } from "metabase/selectors/settings";
 import { getShowMetabaseLinks } from "metabase/selectors/whitelabel";
-import { getIsHosted } from "metabase/setup";
 import { Code } from "metabase/ui";
 
 import type { TipProps as _TipProps } from "./TroubleshootingTip";
