+++ frontend/src/metabase/setup/components/CloudMigrationHelp/CloudMigrationHelp.tsx
@@ -2,9 +2,11 @@ import { t } from "ttag";
 
 import { HelpCard } from "metabase/common/components/HelpCard";
 import { useSelector } from "metabase/redux";
-import { migrateToCloudGuideUrl } from "metabase/selectors/settings";
+import {
+  getIsHosted,
+  migrateToCloudGuideUrl,
+} from "metabase/selectors/settings";
 
-import { getIsHosted } from "../../selectors";
 import { useStep } from "../../useStep";
 import { SetupCardContainer } from "../SetupCardContainer";
