+++ frontend/src/metabase/transforms/selectors.ts
@@ -2,9 +2,12 @@ import { createSelector } from "@reduxjs/toolkit";
 
 import { getPlan } from "metabase/common/utils/plan";
 import type { State } from "metabase/redux/store";
-import { getSetting } from "metabase/selectors/settings";
+import {
+  getIsHosted,
+  getSetting,
+  getTokenFeature,
+} from "metabase/selectors/settings";
 import { getUser, getUserIsAdmin } from "metabase/selectors/user";
-import { getIsHosted, getTokenFeature } from "metabase/setup";
 
 export const canAccessTransforms = (state: State): boolean => {
   if (getUserIsAdmin(state)) {
